health/privnet: let server listen to NodeInternalIPs only
Currently, the health server listens to all IP addresses. However, this
is suboptimal, as it potentially makes the server reachable from the
externally facing networks as well. Let's adapt the default listener
factory to listen to the NodeInternalIP addresses only (both IPv4 and
IPv6 if available), with a fallback to the NodeExternalIP ones. Most
notably, this also matches the symmetric logic to determine the address
to use on the client side (i.e., [types.NewNode]).

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>