privnet/health: relax requirement of same health port across clustermesh
Currently, the health check port must match across the entire Cluster
Mesh. However, this is suboptimal, as different clusters may have
different requirements/constraints (especially considering that it
is allocated in host network), as well as because it is tricky to
subsequently  change without causing a widespread outage.

Let's improve this by having each INB node advertise the selected
port via a CiliumNode annotation. Workload nodes can then leverage
the advertised port when connecting, falling back to the globally
configured one if not specified. To this end, the [tables.INBNode]
object is extended to additionally include the target health port.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>