privnet: Add `privnet/status` command
The `privnet/status` command gives you a summarized status of the
privnet sub-system.
This will directly be used of the `inbadm status` command for the
Isovalent Network Bridge and will act as a basis for the `cilium privnet
status` command, which will collect the status of every node and
summarize it on a cluster level.

Signed-off-by: Fabian Fischer <fabian.fischer@isovalent.com>