helm: Rename Timescape role/bindings for cron TLS cert gen
In Cilium v1.17.7 (42de847053b47204a889a447e72444581b996446),
we fixed an issue where the role referenced in the rolebinding for
Timescape cert generation through cronjob was mistakently set to the
OSS role. Because a RoleBinding roleRef attribute is read-only, this
causes an error on upgrade when Helm tries to update the resource.

To mitigate this, rename both the role and role bindings to force Helm
to delete and then create new resources.

Signed-off-by: Alexandre Barone <abalexandrebarone@gmail.com>