egwha: pass agent policy config by value.
The intention of storing these as pointers was to reduce
copying and memory usage for statedb objects meant to
be immutable.

Unfortunately, this property doesn't really apply to
the composed AgentPolicyConfig type as the matched
endpoints are swapped out - leading to data races
as the "db/show ..." commands don't hold the agent
manager lock.

Instead, this changes the table to a scalar type, the
underlying composed *PolicyConfig is the immutable
part so we can keep that as a reference.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>