egwha: fix issue where policy is never updated.
Previous commits switched the policy config table
to use a scalar type.
This exposed another issue where we were not actually
updating table elements during reconciliation.

Instead we had inadvertadly been using the fact that
the agent policies were stored as references to update
stored elements inside the table when doing regenerate
gateway config.

That behavior was incorrect, and now that this no longer
happens we have to actually re-upsert the row to make the
change.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>