bgp: Introduce EnterpriseRouter interface
As our enterprise implementation grows, it is getting harder to
introduce a new types or struct fields into the OSS types package which
is useless from OSS's standpoint. Also, it takes time to propagate OSS
changes into CEE which significantly affects feature velocity.

This commit introduces EnterpriseRouter interface. This is a superset of
the OSS Router interface and supports "extended" functions. The idea is
we can implement these functions to the OSS GoBGPServer type within
enterprise_xxx.go file and on the enterprise side, we "upgrade" OSS
Router to the EnterpriseRouter type using paramUpgrader. In this way, we
can introduce enterprise-specific logic without modifying OSS.

As a first target, we introduce GetRoutesExtended which returns
Neighbor's Address that advertised the route on top of the existing
GetRoutes response which is required for the BGP Unnumbered support for
BGP Route Import.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>