bgp: Introduce EnterpriseFakeRouter
Introduce a new test facility EnterpriseFakeRouter. This implements an
EnterpriseRouter interface as well as the reset channel feature used in
the BFD reconciler (it replaces the bfdFakeRouer).

We could extend the OSS FakeRouter, but unfortunately that requires
another OSS changes because the FakeRouter is in the OSS types package
and implementing EnterpriseFakeRouter interface introduces circular
import. So, until we fix the OSS side, we keep this copy here.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>