bgp: Support upgrading Router in paramUpgrader
Change the type of the Router field in the EnterpriseBGPInstance to
EnterpriseRouter and support upgrading GoBGPServer and
EnterpriseFakeRouter to the EnterpriseRouter. When we upgrade, we need
to be aware of the underlying Router implementation. This is ugly, but I
didn't come up with the better way to do it without modifying OSS Router
interface.

The rest of the changes are converting the existing tests to use
EnterpriseFakeRouter instead of OSS FakeRouter to make it upgradable.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>