egressgwha: suppress warning when updating non-existent IEGP
Tolerate the situation where the policy resource has been deleted while the
cilium-operator still wants to update it. This should be rare in practice,
but we seem to be hitting it in CI.

Note that this change could go further and also suppress the reconcile
retry for this specific error case. Intentionally not doing this here to
reduce the risk, since we need to backport the change.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>