gha: checkout PR head, instead of merge commit, in ent-build-images-ci
Currently, the "Enterprise Image CI Build" workflow does not specify the
reference to checkout. In turn, that defaults to checking out the merge
commit of the PR onto the target branch, in case of PR, instead of the
head of the PR branch itself. This is a deviation from the behavior of
the upstream "Image CI Build" workflow (which is actually used to build
the non-OLM enterprise images as well). Additionally, it makes the commit
SHA embedded in the images and reported by the various `version` commands
not significant, as pointing to a non-existing commit.

Let's get this fixed explicitly checking out the head of the PR branch,
in case of PR. No difference is introduced in case of push events, as
we still checkout the head of the branch itself.

Fixes: 95d15632960f ("Add CI workflow for building enterprise images")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>