ilb: Dashboards support Node label edge cases
Many of the panels on the ILB dashboards would break under either of the
following conditions:

* The Prometheus Pod had restarted and had a new Pod IP after the
  restart
* The Nodes use a Cloud Controller Manager, and expose a separate
  Prometheus time series for Node names for each of the External IP and
  Internal IP

In either of these cases, the vector multiplication (join) between the
time series and the subquery which extracted the Node names would fail,
resulting in an empty panel (for each of 13 panels) and an error being
shown for all of these panels.

This PR modifies all of those subqueries which extract the Node name to
support those cases by first using only the "InternalIP" version of
those series, and then also picking only one of the "instances" of the
kube-state-metrics pod from which to extract the address (Node name)
vector.

This fixes these panels and restores functionality for the Resource
Utilization dashboard section and the Infrastructure dashboard section.

Signed-off-by: Andrew Sauber <andrew.sauber@isovalent.com>