privnet/hubble: add hubble parser support for privnet flows
The Private Networks feature introduces the concept of "NetIPs", which
are essentially "virtual IPs" that are then mapped to "PIP", which
correspond to actually routable pod IPs. Only the "PIP" make sense for
OSS Cilium and "NetIPs" might overlap with other "PIPs".

Previous commits added support datapath support to add "NetID" to
monitor events, to indicate whether an event contains a PIP or NetIP and
in what Private Network. This commit now extends the OSS Hubble parser
to correctly parse events containing "NetIPs".

To do that, we hook into the OSS Parse at roughly three points:

  1. During monitor event parsing, we extract the new NetIDs and store
     them for later hook points.
  2. After the OSS packet decoder decoded the packet, we translate all
     NetIPs to PIPs, as OSS Hubble only understands PIPs. We store any
     parse NetIP
  3. After the OSS parser returned, we translate all PIPs (if
     applicable) to NetIPs. The user will want to see their "virtual"
     NetIPs in the flows. This applies both to L34 and L7 flows.

Signed-off-by: Fabian Fischer <fabian.fischer@isovalent.com>