enterprise/api/v1: add /network/private/addressing Cilium API endpoint
Add the possibility to query the private networks address and route
configuration via the enterprise OpenAPI swagger API. This information
will be used by the CNI to set up private-networks-enabled endpoints on
CNI ADD in a successive commit.

This commit only contains the model changes and stub handlers. The API
will be implemented in a subsequent change.

Co-authored-by: David Bimmler <david.bimmler@isovalent.com>
Co-authored-by: Fabian Fischer <fabian.fischer@isovalent.com>
Co-authored-by: Harsimran Pabla <hpabla@isovalent.com>
Co-authored-by: Marco Iorio <marco.iorio@isovalent.com>
Co-authored-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>