privnet/bpf: introduce datapath side of the watchdog bpf map
Introduce the datapath side definition of the privnet-watchdog BPF
map, which is intended to be used to detect when the Cilium agent
is down for a prolonged period of time. In turn, this allows stopping
performing L2 announcements towards the external network, as a fail
over may have occurred in the meanwhile, and we may conflict with
the currently active INB, until the local agent is up again.

A similar mechanism already exists today in Cilium, and is provided
via the config_map map. However, that map is not recreated upon agent
restart. Reusing it would be suboptimal in this case, because the other
privnet-related maps are recreated upon agent restart, hence causing a
behavioral mismatch. More specifically, the watchdog map would start
being updated again immediately after an agent restart, but at that
point the other privnet maps would still contain stale data, until the
corresponding endpoints are regenerated. Hence, restarting with the
conflicting L2 announcements, potentially forever if endpoint
regeneration fails.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>