privnet: introduce control-plane side of the watchdog bpf map
Introduce the control-plane side definition of the privnet-watchdog
BPF map, which is intended to be used to detect when the Cilium agent
is down for a prolonged period of time. In turn, this allows stopping
performing L2 announcements towards the external network, as a fail
over may have occurred in the meanwhile, and we may conflict with
the currently active INB, until the local agent is up again.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>