hubble-ee: Allow using service resolver to resolve Timescape grpc endpoint
Introduce a new option in the Hubble Timescape exporter to allow
resolving the Timescape gRPC endpoint IP Address using the Cilium
service resolver based on Cilium's internal service cache.

This will be used by default and will prevent having to set the dnsPolicy
of the cilium-agent to `ClusterFirstWithHostNet`. Changing the dnsPolicy
is required when relying on in-cluster DNS to perform IP Address resolution,
but it can cause issues during cluster bootstrap in certain environments
where the K8s API server host can't be resolved and causes a deadlock
preventing Cilium/the cluster to become ready.

Signed-off-by: Alexandre Barone <abalexandrebarone@gmail.com>