helm: Do not change dnsPolicy when Hubble Timescape exporter enabled
Changing the dnsPolicy is required when relying on in-cluster DNS to perform
IP Address resolution, but it can cause issues during cluster bootstrap in certain
environments where the K8s API server host can't be resolved and causes a
deadlock preventing Cilium/the cluster to become ready.

Instead, we will rely on the Cilium service resolver for in-cluster endpoint
resolution for the Timescape exporter.

Signed-off-by: Alexandre Barone <abalexandrebarone@gmail.com>