privnet/bpf/test: let setup functions return the original error
Currently, the setup functions explicitly return `TEST_ERROR` at the
end, but we only reach this point if the tailcall invocation performed
by the last function invocation failed. Let's instead directly return
the error returned by that function, which is already `TEST_ERROR` if
the tailcall failed, for improved clarity.

While being there, let's also rework the `PRIVNET_ICMP6_NS_SETUP` macro
to be a real function instead, so that the control flow is clearer.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>