privnet: introduce cell to handle watchdog map updates
Introduce the control-plane cell to periodically update the newly
introduced privnet-watchdog map, to signal that the Cilium agent is
still alive. The agent liveness timeout is intended to be configured
externally via the dedicated methods of the [Watchdog] interface, but
can be also overwritten via a hidden flag. Regardless, it is eventually
propagated to the datapath via the watchdog map. The update interval
is instead automatically computed to be the 20% of the timeout.

More in detail, the [Watchdog] interface allows multiple components to
request a given timeout via the [RequestTimeout] method (and eventually
release it via [ReleaseTimeout]). The timeout actually used at any given
point in time is computed as the minimum of all requested timeouts,
falling back to a default of 2 seconds otherwise.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>