ent/alignchecker: correctly validate offsets of struct fields
Currently, all enterprise alignchecker entries are registered via
[registerToCheckSizes], which only validates that the overall size
matches, but not the correspondence of the internal fields. Let's
instead make use of [registerToCheck] to enable the more aggressive
checks, similarly to what we already do upstream.

At the same time, let's fix the align tags of the [EncryptionPolicyKey]
struct, which don't match the corresponding datapath definition, and
are now reported by the check.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>