Introduce a tool to override go.mod
Introduce a new enterprise-specific tool gomod-override. This allows
overriding go.mod file for enterprise-specific changes.

As an initial feature, it supports adding/overriding the `replace`
directive. This allows us to use enterprise-specific fork of the module.
For example, in the OSS, we replace osrg/gobgp to cilium/gobgp, but in
the enterprise, we want to replace it to isovalent/gobgp instead of
cilium/gobgp because we have some patches that only makes sense for
enterprise use cases.

By combining this tool with the usual diff check, we can check whether
the override is properly applied in the CI. It helps backporters to
resolve conflicts and avoid accidental removal of the override.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>