Extend go-mod-tidy merge driver to execute gomod-override
Extend go-mod-tidy merge driver to execute gomod-override before go mod
tidy and go mod vendor. This helps backporters to resolve conflict
automatically even with the override.

This is an enterprise-specific driver, so we define it in the separate
file. This must also work with OSS branch as it is annoying for
backporters to switch between enterprise and OSS merge drivers.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>