egressgwha: restore flaky check in TestEgressCIDRAllocationFloatingIP
5fe8706ec56a ("egressgwha: disable condition check in TestEgressCIDRAllocationFloatingIP")
disabled this check, because the mechanism it used to opt-out Node1 from
active selection as GW was not reliable. makeNodeUnschedulableByTaint()
would internally trigger an update for both the `Node` and `CiliumNode`
resource. If those resources are reconciled separately, the expected
conditions don't apply.

But instead we can use the new Agent-down health status to opt-out Node1
from active selection. This mechanism doesn't suffer from the described
reconciliation race.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>