privnet/cilium-cli: Fix test failure output
When the privnet cilium-cli infrastructure was upstreamed, the `args`
part of `scenario.fail` was changed to be ignored, leading to outputs
such as:

```
Test scenario "curl-client-network-a-to-privnet-vm-net-a1" (ipv4) failed: unexpected curl exit code. got %d want %d"
```

This commit fixes that by re-introducing `msg` as a format string,
so it shows again as:

```
Test scenario "curl-client-network-a-to-privnet-vm-net-a1" (ipv4) failed: unexpected curl exit code. got 0 want 28
```

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>