bgp: Make IPv6 link-local of test iface predictable
In the current BGP Unnumbered test, we are relying on the environment
variable to obtain auto assigned IPv6 link-local address of the test
interfaces. This makes the test case complicated and also the variable
substitutaion of the Hive Script does not play well with table output
comparison (IPv6 address string is variable length, so it's hard to
adjust column header to be aligned with the generated IP).

To simplify that, make the link-local address predictable. We achieve
that by always generating link-local address using EUI64 encoding and
fixed MAC address. In this way, the kernel always generates the same
IPv6 link-local address for the test interfaces so that we can directly
write the address to the test.

Also update the existing ent-unnumbered.txtar to not use variables.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>