privnet/health: don't reduce timeout advertised to the INB
Currently, the workload nodes advertise a reduced timeout (80%) to the
INBs, with the goal of speeding up the detection on the INB side, and
reduce the likelihood of ending up with two INB thinking that they are
both active at the same time. Let's slightly change this, advertising
the full timeout value, and moving the scaling logic to the INB side.
This gives more control to the INB itself, and allows to perform
computations based on the correct timeout value.

We are not concerned of this being a breaking change, because we haven't
released this feature yet.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>