privnet: hook watchdog cell into place, and adapt tests
Register the newly introduced watchdog cell, and adapt the health server
to appropriately invoke the request and release timeout methods whenever
a new workload node is discovered or turns unhealthy.

Concerning the watchdog timeout computation, we are conservative, and
reduce the timeout advertised via the ProbeRequest by the probe interval.
This is to have a reasonable guarantee that it would always expire first,
regardless of the fact that intervals are different, and not synchronized.
The worst case, if we were to use the same timeout, is exemplified by the
following (hpr: health probe, hpt: health probe timeout, alp: agent liveness
probe, alt: agent liveness timeout):

   alp  alp alp alp               alt
    v    v   v   v                 v
  --|--|-|---|---|-|-------|~~~~~~~|
       ^           x down  ^
      hpr         hpr     hpt

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>