contrib/mindful: report divergences of values.yaml.tmpl
Currently, the meditation script has a wildcard exclusion for helm
values files. However, this also matches the upstream template
file, which is not expected to be modified on the enterprise side,
as there's a dedicated template for enterprise-only settings.

Hence, let's tune the exclusions a bit more to report possible
divergences there as well, and make sure they don't go unnoticed.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>