privnet/health: additionally propagate probe interval via the gRPC API
Add the probe interval field to the ProbeRequest message, and adapt the
client and server logic accordingly. The interval value is intended to
be used in subsequent commits to compute the watchdog timeouts.

We are not concerned of this being a breaking change (as we validate
that it is not zero on the server), because we haven't released this
feature yet.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>