ipcache: Fix leak in CIDR metadata consolidation logic
[ upstream commit 50d7f4530e05c5858e51222ed6e54445a98535cf ]

[ Backporter's notes: Resolved conflicts around the Go type
                      of the prefix / cidr. ]

It is possible to leak CIDR entries in the ipcache metadata layer due to
mishandling the ipcache resource for CIDR entries. The CIDR entries were
inserted into the ipcache but in a way where we attempted to optimize
the entries by consolidating all CIDRs that appear in policies into a
single resource per CIDR. See added test for more details on the leak.

Also, the logic checks that it is only a CIDR entry (only
from CIDR statements in policies) before consolidating. This way, we
don't break the FQDN name manager when there is overlap in the CIDRs
because we don't want to consolidate the ipcache metadata resources when
they are of different types.

Fixes: https://github.com/cilium/cilium/pull/39176
Fixes: a3bd1a69b3b262209bfa6a614ab9c86eca4f7d0f ("ipcache: Consolidate prefixes from various resources")
Reported-by: Sebastian Wicki <sebastian@isovalent.com>
Co-authored-by: Odin Ugedal <odin@uged.al>
Co-authored-by: Odin Ugedal <ougedal@palantir.com>
Signed-off-by: Chris Tarazi <chris@isovalent.com>