privnet/health: retry network activations also when explicitly rejected
Currently, workload nodes retry requests to an INB to activate a given
network only in case of generic errors, and not when explicitly denied
by the INB itself, either because the INB thinks that the workload node
is unhealthy, or the requested network cannot be served. Regardless,
that network is immediately marked as Denied on the workload node side,
under the assumption that the node would have anyways eventually received
an event stating that the network is Denied.

However, it is also possible that the activation request gets denied
if the node and the INB don't temporarily agree about each other health
status. While that mismatch is expected to reconverge once receiving
the subsequent health probe (i.e., within one interval), immediately
marking the network as Denied causes that INB to never be considered
again as a valid candidate for that network, until another health
status change occurs.

Let's prevent this by letting also explicit denials to be retried on
the workload node side, before actually marking the network as Denied,
as last resort. This is intended to give enough time for both parties
to agree again on the health state, and either proceed with the activation
or abort it altogether, if the INB is considered Unhealthy.

We may try to better distinguish possibly transient and persistent
failures (e.g., the network is actually not served), and only retry
the former; however, there's very little benefit in doing so, given
that they are expected to occur very rarely anyways, and we'd increase
the risk of ending up in situations like the one addressed by this
commit, causing a persistent failure due to a temporary disagreement.

While being there, let's also add the attempt count to the warning
message, for additional clarity.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>