alignchecker: don't rely on to be removed OSS functions
We relied on OSS functions to register the enterprise structs for
alignchecking. Let's not do this.

This way is actually also more useful for the enterprise side, because
now when the OSS structs are not registered in init, we have a
deterministic ordering between the enterprise and OSS structs, which
would for example allow us to overwrite some of the OSS structs (which
I'll likely need to do in a follow-up PR)

Signed-off-by: Fabian Fischer <fabian.fischer@isovalent.com>