ctmap/gc: fix remainig races related to closing channels.
[ upstream commit 460e9f4d1334a793953728b7a7fda6cee51ed957 ]

We need to correctly synchronize around feedSignals before cleaning up
the feeder channel. As well, we want to avoid races around mutating the
*gc between feeders so we now shallow copy that and pass that separately
to feeders so we can safely swap out the signals channel.

This was evident when rerunning the test a couple times with -race
enabled:

```
go test -v -count=1000 -race -failfast
```

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>