privnet/contrib: introduce the containerlab-based test topology
Introduce the definition of the containerlab-based test topology, which
is used to automatically setup the environment used to test the private
networks functionality. At a high level, it encompasses:

* Overlapping network CIDRs;
* Usage of VLANs, with INBs making use of both VLAN access and trunking;
* Cross-network and DCN routes support.

The testbed includes four user networks (respectively network-{a,b,c,d}),
and network connectivity is provided by a single switch, with VLANs being
used to provide segmentation. network-c and network-d are configured with
overlapping CIDRs, while the other CIDRs are disjoint. One router provides
connectivity between network-a, b, and c and to the Internet, and another
connects network-d to the Internet.

Moreover, the testbed is composed of three KinD clusters, one representing
the main workload cluster (where kubevirt is eventually deployed), while
the other two are Isovalent Network Bridges. The first INB serves network-a
and network-b. The second serves all networks, and is interconnected to
the switch via a trunk interface, with sub-interfaces configured for the
individual VLAN tags.

Finally, a set of containers represent workloads attached to the different
user networks.

Co-authored-by: David Bimmler <david.bimmler@isovalent.com>
Co-authored-by: Fabian Fischer <fabian.fischer@isovalent.com>
Co-authored-by: Harsimran Pabla <hpabla@isovalent.com>
Co-authored-by: Sebastian Wicki <sebastian@isovalent.com>
Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>