privnet/contrib: introduce private-networks specific Makefile
Introduce a Makefile intended to include the private networks specific
targets. For the moment, it includes the targets to setup and destroy
the containerlab-based testbed environment. Additionally, it is also
responsible for populating the information within vcsim, a vSphere Web
Services SDK endpoints for testing purposes. Finally, we also check-in
a python script serving as a basic mock of the MTV inventory.

Co-authored-by: David Bimmler <david.bimmler@isovalent.com>
Co-authored-by: Fabian Fischer <fabian.fischer@isovalent.com>
Co-authored-by: Harsimran Pabla <hpabla@isovalent.com>
Co-authored-by: Sebastian Wicki <sebastian@isovalent.com>
Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>