chore: Build external images on pull requests
To have the OpenShift conformance workflow running on OSS syncs PRs the UBI
versions of the external images need to get built for them as well as the
OpenShift deployment depends on them.
The workflow checks whether an image version is already available and
does not rebuild it if it is.

fixes #9510

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>