privnet/bpf: Add unknown flow policy enforcement
This commit adds policy enforcement for unknown flow via INB.

Policy enforcement is done before NAT and after revNAT, which ensures
that we see the flow with private network IPs (rather than P-IPs). There
are two new BPF maps that are specific to unknown flow.

- There is a map of CT maps. This allows per-privnet connection
  tracking, as every privnet has its own IP space and therefore we need
  to be careful to not conflate connections that are live in different IP
  spaces.
- Rather than looking up the IP identity in IPCache (which is only valid
  for the default network), we look up the identity in the CIDR
  identities map, which contains the identities of unmanaged prefixes that
  are valid across all private networks. This map is populated from
  user-space here https://github.com/isovalent/cilium/pull/9614

Other than that, policy enforcement looks similar to how it looks in the
regular Cilium datpath: We skip policy enforcement for reply traffic of
established connections, we emit a policy verdict event for the first
packet of a connection and we allow the packet to go through in audit
mode.

Unknown flow does not support redirect to L7 proxy, any flows which
would hit the proxy are dropped with "policy denied".

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>