bpf: nodeport: fix daddr in IPIP header
[ upstream commit 4833f172a3be6cfb49f1c4387af74917b15e69d0 ]

With 9530af5ede65 we stopped updating the .daddr in the CT tuple with the
backend address. But the IPIP path was relying on this field for building
its outer header.

Just pull the address straight from the backend struct instead. While at it
also fix up the `backend_local` decision for the case where IPIP skipps
updating the CT tuple.

Fixes: 9530af5ede65 ("bpf: lb: delay tuple->daddr update in lb*_local()")
Signed-off-by: Julian Wiedmann <jwi@isovalent.com>