ilb: change default value handling for accesslog format flags
Currently, the ILB T2 accesslog formats for various accessloggers
(protocols) are configured globally as Operator Go flags. There default
values are defined when defining the Go pflag.

This default values lead to extensive output in the documentation and
in the log of the operator at startup.

Therefore, this commit refactors the default value handling to implicitly
set the default values if the flag isn't provided. With the drawback that
the default pattern is no longer as easily visible to the users (e.g. for
slight modification).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>