bpf: nodeport: fix up IPIP forward path
[ upstream commit 6693e11d50c9490e9c49be497c78929de0a7c09c ]

While the IPIP-style service processing typically skips the DNAT, we still
want to continue in the forwarding path (and not just punt the packet up
to the stack). c171f22725ee broke this when lifting the nodeport_xlate*()
check into the caller.

While at it also fix up the naming of the helper.

Fixes: c171f22725ee ("bpf: lb: split up lb*_local()")
Signed-off-by: Julian Wiedmann <jwi@isovalent.com>