privnet/cli: fix policy cleanup in case of test failure
The blamed commit introduced an explicit step to delete all existing
policies before performing the INB failover tests. However, this turned
out being problematic in case any of the tests executed after applying
the last version of the policies fails (or is canceled), given that the
explicit cleanup step would fail as well due to the context having been
canceled. Yet, the deferred execution of the cleanup, which makes use
of a separate context, would be also skipped, as the map of policies is
incorrectly reset upon the first execution.

Let's get this fixed, skipping the execution of the explicit cleanup in
case the context got canceled, so that the deferred one can take care
of it. Additionally, let's avoid resetting the policies map if any of
the deletions fails, so that the last resort cleanup can still attempt
to fix it.

Fixes: 256ea75390dc ("privnet/cli: explicitly trigger cleanup at the end of the test")
Reported-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>