privnet/cli: let cloud-init manage hosts file for test VMs
Let's configure cloud-init to manage the /etc/hosts file of the client
and echo VMs. This ensures that the file contains an entry for the host
name of the VM itself, and prevents reverse resolution to hang in case
the DNS server is not reachable. For instance, it looks like that the
vm-echo script remained blocked for more than one minute before actually
starting the http server, due to the `socket.getfqdn(host)` call.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>