privnet/cli: leverage readiness probes to wait for test VM readiness
Currently, the privnet tests make use of a custom logic to wait for VM
readiness, manually invoking the hostname command inside the VM itself.
However, this does not actually ensure that, e.g., the echo server has
been started, potentially causing test flakiness. Let's improve this
by making use of the readiness probe support provided by Kubevirt, and
updating the client logic to directly check for the reported readiness
status. The readiness probe is configured to execute a custom script,
which simply checks the hostname in case of client VMs (mimicking the
previous behavior), and additionally validates that the echo server
is reachable on the echo VMs. We cannot make use of HTTP probes at
the moment, instead, as not working in the context of private-networks.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>