oper/bgp: Fix crash by empty bgp-secrets-namespace
If the bgp-secrets-namespace setting is not provided, the
secretResource passed to the peerConfigStatusReconciler is nil,
and we should not attempt to initiate the store for it.

This can happen when BGP CP is enabled via cilium CLI
(or manually via config), as it is otherwise defaulted
in the Helm chart.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>