cmapisrv/test: miscellaneous fixes to the ciliumidentities script test
[ upstream commit 6a4142ef9a0d022a34e26e262b3c845726cce476 ]

It appears that the ciliumidentities.txtar test is affected by two
issues. First, the update performed at line 19 is ineffective, because
none of the security labels contain the 'foo' token. Second, the
namespace labels of the second identity are inconsistent. Let's get
them fixed, and update the namespace labels to be different between
the two identities.

Fixes: 1d8c53fac4bf ("test/cmapisrv: add initial battery of script tests")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Antony Reynaud <antony.reynaud@isovalent.com>