enterprise/maps: Add cilium_privnet_cidr_identity BPF map
This commit adds the user-space pluming for the CIDR identity BPF map.
This LPM map associates the identity of a CIDR with a particular
security identity. The map is global, i.e. shared across all private
networks, as the contained CIDRs are assumed to not belong to any
Cilium-managed endpoints.

As with other private networks map, the BPF map is re-created every time
the agent is restarted. However, since we want to preserve the identity
mapping across restarts, the map contents of the old map is read out
before it is re-created. This allows the map reconciler (added in a
subsequent commit) to re-allocate the same numeric identity.

The new constructor is not yet added to Hive, as the downstream consumer
of the observable is not yet there either.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>