privnet: Add CIDR identity and CIDR metadata tables
This commit adds two tables, which are siblings of each other:

The first one is the CIDR metadata table, which tracks the metadata
(i.e. `cidr` and `cidrgroup` labels) associated with a particular CIDR
and their respective owner.

The second one is the CIDR identity table, which then maps each CIDR to
a concrete Cilium security identity.

For every entry in the CIDR metadata table, we expect an entry in the
CIDR identity table to be created eventually.

The reason for having two tables is that it makes the reconciler code
much simpler, as allocation is a two-phase process: First, we need to
collect all metadata from all owners, and only then do we want to
allocate new identities.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>