privnet: Add CIDR identity table reconciler
This adds the second of the three CIDR identities reconcilers.

This second reconciler watches the CIDR metadata table (populated in the
last commit), and allocates identities based on metadata associated with
each CIDR. It also marks new or changed entries as pending, which will
cause the BPF reconciler (added in the next commit) to write each
identity into the actual BPF map.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>