privnet/policy: Collect CIDR labels
This commit add the necessary code to collect CIDR labels from policies.
We do this by intercepting the policy import, rather than IPCache,
because this allows us to only collect CIDRs actually found in policies.
This means that CIDR labels that are derived from other sources (such as
e.g. headless services or node IPs in `policy-cidr-match-mode: nodes`),
which only have meaning in the default network, are not considered for
private network unknown flow.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>