privnet/policy: Rewrite endpoint selectors
This commit adds the logic to rewrite endpoint selectors in network
policies. The rewrite function ensures that policies not having any
requirement on a private network will only select endpoints in the
default network.

If a subject selector on the other hand does require a certain private
network label, then this requirement is copied into all peer selectors
as well (ensuring that we do not select endpoints in other networks,
which are unreachable and thus do not need to be part of the policy
map).

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>